+++ src/DataGrid.tsx
@@ -103,10 +103,6 @@ export interface DataGridProps<R, K extends keyof R> {
   /**
    * Event props
    */
-  /** Function called whenever a row is clicked */
-  onRowClick?(rowIdx: number, row: R, column: CalculatedColumn<R>): void;
-  /** Function called whenever a row is double clicked */
-  onRowDoubleClick?(rowIdx: number, row: R, column: CalculatedColumn<R>): void;
   /** Function called whenever a key is pressed down */
   onGridKeyDown?(event: React.KeyboardEvent<HTMLDivElement>): void;
   /** Function called whenever a key is released */
@@ -327,8 +323,6 @@ function DataGrid<R, K extends keyof R>({
               onRowsUpdate={handleRowUpdate}
               onSelectedCellChange={props.onSelectedCellChange}
               onSelectedCellRangeChange={props.onSelectedCellRangeChange}
-              onRowClick={props.onRowClick}
-              onRowDoubleClick={props.onRowDoubleClick}
               onRowExpandToggle={props.onRowExpandToggle}
             />
           )}